Moved the AtlasSerializer class
