+++ pyatlas/doc/atlas.html
@@ -13,6 +13,19 @@
     <p><tt>The&nbsp;pyatlas&nbsp;<a href="#Atlas">Atlas</a>&nbsp;implementation</tt></p>
 <p>
 <table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
+<tr bgcolor="#aa55cc">
+<td colspan=3 valign=bottom>&nbsp;<br>
+<font color="#ffffff" face="helvetica, arial"><big><strong>Modules</strong></big></font></td></tr>
+    
+<tr><td bgcolor="#aa55cc"><tt>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</tt></td><td>&nbsp;</td>
+<td width="100%"><table width="100%" summary="list"><tr><td width="25%" valign=top><a href="atlas_metadata.html">atlas_metadata</a><br>
+<a href="autogen.html">autogen</a><br>
+</td><td width="25%" valign=top><a href="integer_dictionary.html">integer_dictionary</a><br>
+<a href="packed_tag_store.html">packed_tag_store</a><br>
+</td><td width="25%" valign=top><a href="polyline.html">polyline</a><br>
+<a href="zipfile.html">zipfile</a><br>
+</td><td width="25%" valign=top></td></tr></table></td></tr></table><p>
+<table width="100%" cellspacing=0 cellpadding=2 border=0 summary="section">
 <tr bgcolor="#ee77aa">
 <td colspan=3 valign=bottom>&nbsp;<br>
 <font color="#ffffff" face="helvetica, arial"><big><strong>Classes</strong></big></font></td></tr>
